DjangoTrainProject/requirements.txt
tensorflow>=2.0
keras>=2.3.0
matplotlib>=3.0
split-folders>=0.5.0
scikit-learn>=0.24.0
mlxtend>=0.19.0
matplotlib>=3.0
Pillow==9.3.0  
django==4.2  
